+++ gwcelery/tasks/external_triggers.py
@@ -146,8 +146,7 @@ def handle_grb_lvalert(alert):
         elif alert['alert_type'] == 'label_added' and \
                 alert['data']['name'] == 'EM_COINC':
             ligo_fermi_skymaps.create_combined_skymap(graceid).delay()
-            raven.calculate_spacetime_coincidence_far(graceid,
-                                                      group).delay()
+            raven.calculate_coincidence_far(graceid, group).delay()
 
 
 @lvalert.handler('superevent',